Incomplete - issue 35: Profile's picture http://code.google.com/p/scu-smp/issues/detail?id=35
* Refactoring

+++ smp-php/smp/command/usermentor/EditUserCommand.php
@@ -13,7 +13,8 @@ require_once('Image/Transform.php');
 class smp_command_usermentor_EditUserCommand extends smp_command_Command {
 
 	function doExecute(smp_controller_Request $request) {
-		$userId = $request->getProperty('userId');
+		$currentUser = smp_base_SessionRegistry::getUser();
+		$userId = $currentUser->getId();
 		
 		$userService = new smp_service_UserService();
 		
@@ -28,7 +29,7 @@ class smp_command_usermentor_EditUserCommand extends smp_command_Command {
 			$validator->checkWithRegex("scuEmail", "SCU-Email is not valid SCU account. e.g. fbar12@scu.edu.au", "/^[a-z0-9A-Z_\\.\\-]+@scu.edu.au$/");
 			
 			if ($validator->isValid()) {
-				$user = $userService->find($validator->getProperty('userId'));
+				$user = $userService->find($userId);
 				$user->setScuEmail($validator->getProperty('scuEmail'));
 				$user->setPassword(md5($validator->getProperty('password')));
 				
@@ -46,7 +47,7 @@ class smp_command_usermentor_EditUserCommand extends smp_command_Command {
 					$imgTrans->scaleByXY(100,100);
 					$imgTrans->save(Constants::IMAGE_UPLOAD_DIR."_thb_".$picture);
 					// First remove old picture from image directory
-					if (file_exists(Constants::IMAGE_UPLOAD_DIR.$user->getPicture())) {
+					if (file_exists(Constants::IMAGE_UPLOAD_DIR.$user->getPicture()) && !is_null($user->getPicture())) {
 						unlink(Constants::IMAGE_UPLOAD_DIR.$user->getPicture());
 						unlink(Constants::IMAGE_UPLOAD_DIR."_thb_".$user->getPicture());
 					}
